requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==1.5.4
billiard==4.1.0
cachetools==4.1.1
celery==5.3.4
certifi==2020.4.5.1
cffi==1.16.0
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorclass==2.2.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
cssbeautifier==1.14.9
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
dj-database-url==0.5.0
Django==5.0.4
django-admin-rangefilter==0.11.1
django-autoslug==1.9.9
django-background-tasks==1.2.5
django-cities-light==3.9.2
django-compat==1.0.15
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-datatables-view==1.20.0
django-dob-widget==1.2.0
django-filter==23.5
django-htmx==1.16.0
django-import-export==3.2.0
django-jazzmin==3.0.0
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
django-shortuuidfield==0.1.3
django-tenant-schemas==1.12.0
django-tenants==3.6.1
django-upload-validator==1.1.6
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djlint==1.34.0
docopt==0.6.2
drf-spectacular==0.27.0
drf-yasg==1.21.7
EditorConfig==0.12.3
et-xmlfile==1.1.0
facebook-sdk==3.1.0
Faker==4.1.0
filelock==3.12.2
future==0.18.2
google-api-core==1.22.2
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.9
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsbeautifier==1.14.9
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector==2.2.9
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
pip-upgrader==1.4.15
pipenv==2023.8.20
platformdirs==3.10.0
progressbar2==4.2.0
prompt-toolkit==3.0.39
protobuf==3.13.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.8
python-magic==0.4.27
python-twitter==3.5
python-utils==3.7.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
regex==2023.10.3
reportlab==4.0.4
requests==2.23.0
requests-oauthlib==1.3.0
rpds-py==0.15.2
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==24.1.0
shortuuid==1.0.11
six==1.16.0
social-auth-core==4.5.0
sqlparse==0.4.4
tablib==3.5.0
terminaltables==3.1.10
text-unidecode==1.3
toml==0.10.1
tqdm==4.66.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.25.9
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.3
